requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-surprise>=1.1.0
torch>=1.12.0
torchvision>=0.13.0
requests>=2.28.0
tqdm>=4.64.0
flask>=2.0.0 